ML backend/pyproject.toml
[tool.poetry]
name = "ml backend"
version = "0.1.0"
description = ""
authors = ["vivek pawar <vivekpawarvp07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
torch = "^2.0.0"
transformers = "^4.30.0"
textblob = "^0.17.1"
sentence-transformers = "^2.2.2"
Pillow = "^10.0.0"
scikit-learn = "^1.3.0"
scipy = "^1.11.0"
numpy = "^1.25.0"
sentencepiece = "^0.2.0"
python-multipart = "^0.0.12"
nudenet = "^3.4.2"
langchain-groq = "^0.2.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.5"
pandas = "^2.2.3"
keras = "^3.6.0"
tensorflow = "^2.18.0"
xgboost = "^2.1.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
statsmodels = "^0.14.4"
lightgbm = "^4.5.0"
pytorch-lightning = "^2.4.0"
fastai = "^2.7.18"
catboost = "^1.2.7"
gensim = "^4.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"